05/05

+++ server/routes/answers.js
@@ -1,3 +1,4 @@
+// Routing to /posts/answers
 const express = require('express');
 const router = express.Router();
 
@@ -10,15 +11,12 @@ const auth = require('../middleware/auth');
 
 router.get('/:qid', async (req, res) => {
   try {
-    const foundQuestion = await Questions.findById(req.params.qid).exec();
-    const foundAnswer = await Answers.find({ _id: { $in: foundQuestion.answers } }).sort({ ans_date_time: -1 });
-    switch (true) {
-      case foundAnswer && foundAnswer.length > 0:
-        res.send(foundAnswer);
-        break;
-      default:
-        res.status(404).send('Answer not found');
-        break;
+    const question = await Questions.findById(req.params.qid).exec();
+    const answer = await Answers.find({ _id: { $in: question.answers } }).sort({ ans_date_time: -1 });
+    if (answer) {
+      res.send(answer);
+    } else {
+      res.status(404).send('Answer not found');
     }
   } catch (err) {
     console.error(err);
@@ -28,15 +26,12 @@ router.get('/:qid', async (req, res) => {
 
 router.get('/comments/:answer_id', async (req, res) => {
   try {
-    const foundAnswer = await Answers.findById(req.params.answer_id).exec();
-    const foundComment = await Comments.find({ _id: { $in: foundAnswer.comments } }).sort({ com_date_time: -1 });
-    switch (true) {
-      case foundComment && foundComment.length > 0:
-        res.send(foundComment);
-        break;
-      default:
-        res.status(404).send('Comment not found');
-        break;
+    const answer = await Answers.findById(req.params.answer_id).exec();
+    const comment = await Comments.find({ _id: { $in: answer.comments } }).sort({ com_date_time: -1 });
+    if (comment) {
+      res.send(comment);
+    } else {
+      res.status(404).send('Answer not found');
     }
   } catch (err) {
     console.error(err);
@@ -47,8 +42,8 @@ router.get('/comments/:answer_id', async (req, res) => {
 router.get('/getQuestionsAnswered/:user_id', async (req, res) => {
   try {
     const answersByUser = await Answers.find({ ans_by: req.params.user_id }).exec();
-    const foundQuestions = await Questions.find({ answers: { $in: answersByUser } }).sort({ ask_date_time: -1 });
-    res.send(foundQuestions);
+    const questions = await Questions.find({ answers: { $in: answersByUser } }).sort({ ask_date_time: -1 });
+    res.send(questions);
   } catch (err) {
     console.error(err);
     res.status(500).send('Internal Server Error');
@@ -58,26 +53,26 @@ router.get('/getQuestionsAnswered/:user_id', async (req, res) => {
 router.use(auth); // ANYTHING BELOW THIS WILL REQUIRE AUTHENTICATION
 
 router.post('/answerQuestion', async (req, res) => {
-  let newAnswerData = req.body;
-  newAnswerData.ans_by = req.session.user.userId; // do not trust the client to send the user id via post request
+  let newAnswerInput = req.body;
+  newAnswerInput.ans_by = req.session.user.userId; // do not trust the client to send the user id via post request
   try {
     const newAnswer = new Answers({
-      text: newAnswerData.text,
-      ans_by: newAnswerData.ans_by,
+      text: newAnswerInput.text,
+      ans_by: newAnswerInput.ans_by,
     });
     await newAnswer.save();
 
-    if (!newAnswerData.qid) {
+    if (!newAnswerInput.qid) {
       res.status(400).send('Missing qid parameter');
       return;
     }
 
-    const foundQuestion = await Questions.findById(newAnswerData.qid).exec();
+    const question = await Questions.findById(newAnswerInput.qid).exec();
 
-    if (foundQuestion) {
-      foundQuestion.answers.push(newAnswer._id);
-      await foundQuestion.save();
-      res.send(foundQuestion);
+    if (question) {
+      question.answers.push(newAnswer._id);
+      await question.save();
+      res.send(question);
     } else {
       res.status(404).send('Question not found');
     }
@@ -89,25 +84,22 @@ router.post('/answerQuestion', async (req, res) => {
 
 router.delete('/deleteAnswer/:answer_id', async (req, res) => {
   try {
-    const foundAnswer = await Answers.findById(req.params.answer_id).exec();
-    switch (true) {
-      case foundAnswer:
-        const foundQuestion = await Questions.find({ answers: { $in: foundAnswer } }).exec();
-        if (foundQuestion) {
-          foundQuestion[0].answers.pull(foundAnswer._id);
-          await foundQuestion[0].save();
-        }
+    const answer = await Answers.findById(req.params.answer_id).exec();
+    if (answer) {
+      const question = await Questions.find({ answers: { $in: answer } }).exec();
+      if (question) {
+        question[0].answers.pull(answer._id);
+        await question[0].save();
+      }
 
-        foundAnswer.comments.forEach(async (comment) => {
-          await Comments.deleteOne({ _id: comment._id }).exec();
-        });
+      answer.comments.forEach(async (comment) => {
+        await Comments.deleteOne({ _id: comment._id }).exec();
+      });
 
-        await Answers.deleteOne({ _id: req.params.answer_id }).exec();
-        res.send('Success');
-        break;
-      default:
-        res.status(404).send('Answer not found');
-        break;
+      await Answers.deleteOne({ _id: req.params.answer_id }).exec();
+      res.send('success');
+    } else {
+      res.status(404).send('Answer not found');
     }
   } catch (err) {
     console.error(err);
@@ -117,15 +109,12 @@ router.delete('/deleteAnswer/:answer_id', async (req, res) => {
 
 router.put('/editAnswer/:answer_id', async (req, res) => {
   try {
-    const foundAnswer = await Answers.findById(req.params.answer_id).exec();
-    switch (true) {
-      case foundAnswer:
-        const result = await Answers.updateOne({ _id: req.params.answer_id }, { $set: { text: req.body.text } }).exec();
-        res.send(result);
-        break;
-      default:
-        res.status(404).send('Answer not found');
-        break;
+    const answer = await Answers.findById(req.params.answer_id).exec();
+    if (answer) {
+      const result = await Answers.updateOne({ _id: req.params.answer_id }, { $set: { text: req.body.text } }).exec();
+      res.send(result);
+    } else {
+      res.status(404).send('Answer not found');
     }
   } catch (err) {
     console.error(err);
@@ -134,105 +123,96 @@ router.put('/editAnswer/:answer_id', async (req, res) => {
 });
 
 router.patch('/incrementVotes/:answer/:userVoted', async (req, res) => {
-  const foundAnswer = await Answers.findById(req.params.answer).exec();
+  const answer = await Answers.findById(req.params.answer).exec();
   let updateUserReputation = 0;
-  switch (true) {
-    case foundAnswer:
-      let voterObj = foundAnswer.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
-      if (voterObj.length > 0) {
-        let userVoted = voterObj[0].userVoted;
-        let currentDirection = voterObj[0].direction;
-        if (currentDirection === -1) {
-          foundAnswer.votes += 1;
-          updateUserReputation = 10;
-        } else if (currentDirection === 0) {
-          foundAnswer.votes += 1;
-          updateUserReputation = 5;
-        }
-        let direction = Math.min(currentDirection + 1, 1);
-        const objIndex = foundAnswer.voters.findIndex((obj) => obj.userVoted == userVoted);
-        foundAnswer.voters[objIndex].direction = direction;
-      } else {
-        foundAnswer.votes += 1;
+  if (answer) {
+    let voterObj = answer.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
+    if (voterObj.length > 0) {
+      let userVoted = voterObj[0].userVoted;
+      let currentDirection = voterObj[0].direction;
+      if (currentDirection === -1) {
+        answer.votes += 1;
+        updateUserReputation = 10;
+      } else if (currentDirection === 0) {
+        answer.votes += 1;
         updateUserReputation = 5;
-        foundAnswer.voters.push({
-          userVoted: req.params.userVoted,
-          direction: 1,
-        });
       }
-      await foundAnswer.save();
-      let userToUpdate = await Users.findOne({ _id: foundAnswer.ans_by }).exec();
-      userToUpdate.reputation += updateUserReputation;
-      await userToUpdate.save();
-      res.status(200).send(foundAnswer);
-      break;
-    default:
-      res.status(404).send('Answer not found');
-      break;
+      let direction = Math.min(currentDirection + 1, 1);
+      const objIndex = answer.voters.findIndex((obj) => obj.userVoted == userVoted);
+      answer.voters[objIndex].direction = direction;
+    } else {
+      answer.votes += 1;
+      updateUserReputation = 5;
+      answer.voters.push({
+        userVoted: req.params.userVoted,
+        direction: 1,
+      });
+    }
+    await answer.save();
+    let userToUpdate = await Users.findOne({ _id: answer.ans_by }).exec();
+    userToUpdate.reputation += updateUserReputation;
+    await userToUpdate.save();
+    res.status(200).send(answer);
+  } else {
+    res.status(404).send('Answer not found');
   }
 });
 
 router.patch('/comments/incrementVotes/:comment/:userVoted', async (req, res) => {
-  const foundComment = await Comments.findById(req.params.comment).exec();
+  const comment = await Comments.findById(req.params.comment).exec();
   let updateUserReputation = 0;
-  switch (true) {
-    case foundComment:
-      let voterObj = foundComment.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
-      if (voterObj.length === 0) {
-        foundComment.votes += 1;
-        updateUserReputation = 5;
-        foundComment.voters.push({
-          userVoted: req.params.userVoted,
-        });
-      }
-      await foundComment.save();
-      let userToUpdate = await Users.findOne({ _id: foundComment.com_by }).exec();
-      userToUpdate.reputation += updateUserReputation;
-      await userToUpdate.save();
-      res.status(200).send(foundComment);
-      break;
-    default:
-      res.status(404).send('Comment not found');
-      break;
+  if (comment) {
+    let voterObj = comment.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
+    if (voterObj.length === 0) {
+      comment.votes += 1;
+      updateUserReputation = 5;
+      comment.voters.push({
+        userVoted: req.params.userVoted,
+      });
+    }
+    await comment.save();
+    let userToUpdate = await Users.findOne({ _id: comment.com_by }).exec();
+    userToUpdate.reputation += updateUserReputation;
+    await userToUpdate.save();
+    res.status(200).send(comment);
+  } else {
+    res.status(404).send('Question not found');
   }
 });
 
 router.patch('/decrementVotes/:answer/:userVoted', async (req, res) => {
-  const foundAnswer = await Answers.findById(req.params.answer).exec();
+  const answer = await Answers.findById(req.params.answer).exec();
   let updateUserReputation = 0;
-  switch (true) {
-    case foundAnswer:
-      let voterObj = foundAnswer.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
-      if (voterObj.length > 0) {
-        let userVoted = voterObj[0].userVoted;
-        let currentDirection = voterObj[0].direction;
-        if (currentDirection === 1) {
-          foundAnswer.votes -= 1;
-          updateUserReputation = 5;
-        } else if (currentDirection === 0) {
-          foundAnswer.votes -= 1;
-          updateUserReputation = 10;
-        }
-        let direction = Math.max(currentDirection - 1, -1);
-        const objIndex = foundAnswer.voters.findIndex((obj) => obj.userVoted == userVoted);
-        foundAnswer.voters[objIndex].direction = direction;
-      } else {
-        foundAnswer.votes -= 1;
+  if (answer) {
+    let voterObj = answer.voters.filter((voter) => voter.userVoted.toString() === req.params.userVoted);
+    if (voterObj.length > 0) {
+      let userVoted = voterObj[0].userVoted;
+      let currentDirection = voterObj[0].direction;
+      if (currentDirection === 1) {
+        answer.votes -= 1;
+        updateUserReputation = 5;
+      } else if (currentDirection === 0) {
+        answer.votes -= 1;
         updateUserReputation = 10;
-        foundAnswer.voters.push({
-          userVoted: req.params.userVoted,
-          direction: -1,
-        });
       }
-      await foundAnswer.save();
-      let userToUpdate = await Users.findOne({ _id: foundAnswer.ans_by }).exec();
-      userToUpdate.reputation -= updateUserReputation;
-      await userToUpdate.save();
-      res.status(200).send(foundAnswer);
-      break;
-    default:
-      res.status(404).send('Answer not found');
-      break;
+      let direction = Math.max(currentDirection - 1, -1);
+      const objIndex = answer.voters.findIndex((obj) => obj.userVoted == userVoted);
+      answer.voters[objIndex].direction = direction;
+    } else {
+      answer.votes -= 1;
+      updateUserReputation = 10;
+      answer.voters.push({
+        userVoted: req.params.userVoted,
+        direction: -1,
+      });
+    }
+    await answer.save();
+    let userToUpdate = await Users.findOne({ _id: answer.ans_by }).exec();
+    userToUpdate.reputation -= updateUserReputation;
+    await userToUpdate.save();
+    res.status(200).send(answer);
+  } else {
+    res.status(404).send('Answer not found');
   }
 });
